# Speed up equal files check